requirements.txt
# Core dependencies （numpy must be coordinate with imgaug)
numpy==1.24.4
opencv-python==4.12.0.88

# Deep Learning Framework (CUDA 11.8)
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

# Data augmentation
imgaug==0.4.0

# Video processing
ffmpeg-python==0.2.0

# Visualization and logging
tensorboard==2.20.0

# Additional suggested dependencies (optional)
# pandas>=1.3.0
# matplotlib>=3.3.0
# tqdm>=4.60.0
# Pillow>=9.0.0